@@ -1,12 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
 import './index.css';
 import 'reactUI/dist/main.css'
 import App from './App';
 
-ReactDOM.render(
-  <React.StrictMode>
+/** Rendering the application */
+const root = ReactDOM.createRoot(document.getElementById("root") as Element);
+export default root.render(
+  //<React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  //</React.StrictMode>
 );
+
